Move execution_time to Test and merge util functions
