+++ tests/data_for_tests/edge_factory.py
@@ -1,4 +1,4 @@
-from src.im2im import are_both_same_data_repr
+from src.im2im.util import are_both_same_data_repr
 
 
 def numpy_rgb_to_bgr(source_metadata, target_metadata):
